Removed Async calls, as it added too much complexity. And my computer can not handle the amount of async calls.
Added backfill handling.
CSV data is older than the expected 9-30-2024, so CSV data had to be manipulated to 12-2024 to allow backfill propagation.

Need to handle customer creation and handling of customers. Currently every time a customer is called from the CSV the customer creation method is called. But need to handle duplication to not create mutiples of the same customer. Instead need checking of a dictionary list that handles the names of the customer's "account_id".

Added a time delay for the backfill as the dates are all on the same day and this throws a 400 code because the dates cannot be the same. Added a simple 9 day delay to the backfill end date. Can be adjusted as needed.